changed score back to int (instead of double)
removed hammer name and actionbar message from config and added them to the Languages

(Finally) dissected the Game class into it's own folder and classes

added language messages for Reward classes

added item_displays and interactions for interactable rewards.

fixed actionbar to be more precise on score (before it updated once per half a second, now it updates every hit made)

added Miscellaneous class for color and getSkull

+++ src/main/java/whackamole/whackamole/Commands.java
@@ -8,6 +8,8 @@ import org.bukkit.ChatColor;
 import org.bukkit.block.BlockFace;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
+import whackamole.whackamole.GS.Game;
+import whackamole.whackamole.GS.GamesManager;
 import whackamole.whackamole.RS.RewardsManager;
 import whackamole.whackamole.Utils.Econ;
 import whackamole.whackamole.Utils.Logger;
@@ -229,7 +231,7 @@ public class Commands {
                                     player.sendMessage(Config.AppConfig.PREFIX + Translator.COMMANDS_SETTINGS_MAXMISSED_SUCCESS.Format(args.get(2).toString()));
                                 }
                                 case SCOREPOINTS -> {
-                                    game.setPointsPerKill (Double.parseDouble((String) args.get(2)));
+                                    game.setPointsPerKill (Integer.parseInt((String) args.get(2)));
                                     player.sendMessage(Config.AppConfig.PREFIX + Translator.COMMANDS_SETTINGS_SCOREPOINTS_SUCCESS.Format(args.get(2).toString()));
                                 }
                                 case SPAWNTIMER -> {
